refactor: fix circular dependencies
