+++ components/checkbox/src/checkbox-field/checkbox-field.js
@@ -3,7 +3,7 @@ import { Field } from '@dhis2-ui/field'
 import { Required } from '@dhis2-ui/required'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { Checkbox } from '../index.js'
+import { Checkbox } from '../checkbox'
 
 const AddRequired = ({ label, required, dataTest }) => (
     <React.Fragment>
