+++ components/file-input/src/file-input-field-with-list/file-input-field-with-list.js
@@ -1,7 +1,7 @@
 import { sharedPropTypes } from '@dhis2/ui-constants'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { FileInputField } from '../index.js'
+import { FileInputField } from '../file-input-field'
 import i18n from '../locales/index.js'
 import { FileListItemWithRemove } from './file-list-item-with-remove.js'
