+++ components/file-input/src/file-input-field/file-input-field.js
@@ -3,7 +3,8 @@ import { Field } from '@dhis2-ui/field'
 import { Label } from '@dhis2-ui/label'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { FileInput, FileList, FileListPlaceholder } from '../index.js'
+import { FileInput } from '../file-input'
+import { FileList, FileListPlaceholder } from '../file-list'
 import i18n from '../locales/index.js'
 
 // TODO: i18n
