+++ components/select/src/single-select-field/single-select-field.js
@@ -3,7 +3,7 @@ import { Box } from '@dhis2-ui/box'
 import { Field } from '@dhis2-ui/field'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { SingleSelect } from '../index.js'
+import { SingleSelect } from '../single-select'
 import i18n from '../locales/index.js'
 
 // TODO: translate
