+++ components/switch/src/switch-field/switch-field.js
@@ -3,7 +3,7 @@ import { Field } from '@dhis2-ui/field'
 import { Required } from '@dhis2-ui/required'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { Switch } from '../index.js'
+import { Switch } from '../switch'
 
 const AddRequired = ({ label, required, dataTest }) => (
     <React.Fragment>
